ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
ortools==9.10.4067
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.4.0
